requirements.txt
torch>=2.0.0
pymatgen>=2023.0.0
h5py>=3.8.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
matplotlib>=3.7.0
pytest>=7.3.0
hypothesis>=6.75.0
